+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add copy shortcut in Resource
+
 ## [0.0.35] - 2025-11-07
 
 ### Added